// frontend/src/app/core/enum/Role&PositionType.enum.ts
// RoleType.enum.ts
export enum RoleType {
  ADMIN = 'admin',
  STUDENT = 'étudiant',
  TEACHER = 'enseignant',
  INDUSTRIAL_TUTOR = 'tuteur_industriel'
}

// PositionType.enum.ts
export enum PositionType {
  ASSISTANT_PROF = 'professeur_assistant',
  PROFESSOR = 'Professeur',
  DEPARTMENT_HEAD = 'Chef_departement',
  HEADMASTER = 'Directeur'
}